+++ spring-cloud-starter-tencent-polaris-router/src/test/java/com/tencent/cloud/polaris/router/PolarisRouterContextTest.java
@@ -20,7 +20,9 @@ package com.tencent.cloud.polaris.router;
 
 import java.util.HashMap;
 import java.util.Map;
+import java.util.Set;
 
+import com.google.common.collect.Sets;
 import org.junit.Assert;
 import org.junit.Test;
 
@@ -38,27 +40,76 @@ public class PolarisRouterContextTest {
 		labels.put("k2", "v2");
 
 		PolarisRouterContext routerContext = new PolarisRouterContext();
-		routerContext.setLabels(PolarisRouterContext.RULE_ROUTER_LABELS, labels);
+		routerContext.putLabels(PolarisRouterContext.ROUTER_LABELS, labels);
 
 		Assert.assertEquals(0, routerContext.getLabels(PolarisRouterContext.TRANSITIVE_LABELS).size());
-		Assert.assertEquals(2, routerContext.getLabels(PolarisRouterContext.RULE_ROUTER_LABELS).size());
-		Assert.assertEquals("v1", routerContext.getLabels(PolarisRouterContext.RULE_ROUTER_LABELS).get("k1"));
-		Assert.assertEquals("v2", routerContext.getLabels(PolarisRouterContext.RULE_ROUTER_LABELS).get("k2"));
-		Assert.assertNull(routerContext.getLabels(PolarisRouterContext.RULE_ROUTER_LABELS).get("k3"));
+		Assert.assertEquals(2, routerContext.getLabels(PolarisRouterContext.ROUTER_LABELS).size());
+		Assert.assertEquals("v1", routerContext.getLabels(PolarisRouterContext.ROUTER_LABELS).get("k1"));
+		Assert.assertEquals("v2", routerContext.getLabels(PolarisRouterContext.ROUTER_LABELS).get("k2"));
+		Assert.assertNull(routerContext.getLabels(PolarisRouterContext.ROUTER_LABELS).get("k3"));
 	}
 
 	@Test
 	public void testSetNull() {
 		PolarisRouterContext routerContext = new PolarisRouterContext();
-		routerContext.setLabels(PolarisRouterContext.RULE_ROUTER_LABELS, null);
+		routerContext.putLabels(PolarisRouterContext.ROUTER_LABELS, null);
 		Assert.assertEquals(0, routerContext.getLabels(PolarisRouterContext.TRANSITIVE_LABELS).size());
-		Assert.assertEquals(0, routerContext.getLabels(PolarisRouterContext.RULE_ROUTER_LABELS).size());
+		Assert.assertEquals(0, routerContext.getLabels(PolarisRouterContext.ROUTER_LABELS).size());
 	}
 
 	@Test
 	public void testGetEmptyRouterContext() {
 		PolarisRouterContext routerContext = new PolarisRouterContext();
 		Assert.assertEquals(0, routerContext.getLabels(PolarisRouterContext.TRANSITIVE_LABELS).size());
-		Assert.assertEquals(0, routerContext.getLabels(PolarisRouterContext.RULE_ROUTER_LABELS).size());
+		Assert.assertEquals(0, routerContext.getLabels(PolarisRouterContext.ROUTER_LABELS).size());
+	}
+
+	@Test
+	public void testGetLabelByKeys() {
+		Map<String, String> labels = new HashMap<>();
+		labels.put("k1", "v1");
+		labels.put("k2", "v2");
+		labels.put("k3", "v3");
+
+		PolarisRouterContext routerContext = new PolarisRouterContext();
+		routerContext.putLabels(PolarisRouterContext.ROUTER_LABELS, labels);
+
+		Map<String, String> resolvedLabels = routerContext.getLabels(PolarisRouterContext.ROUTER_LABELS,
+				Sets.newHashSet("k1", "k2", "k4"));
+
+		Assert.assertEquals(2, resolvedLabels.size());
+		Assert.assertEquals("v1", resolvedLabels.get("k1"));
+		Assert.assertEquals("v2", resolvedLabels.get("k2"));
+	}
+
+	@Test
+	public void testGetLabel() {
+		Map<String, String> labels = new HashMap<>();
+		labels.put("k1", "v1");
+		labels.put("k2", "v2");
+		labels.put("k3", "v3");
+
+		PolarisRouterContext routerContext = new PolarisRouterContext();
+		routerContext.putLabels(PolarisRouterContext.ROUTER_LABELS, labels);
+
+		String resolvedLabel = routerContext.getLabel("k1");
+
+		Assert.assertEquals("v1", resolvedLabel);
+	}
+
+	@Test
+	public void testGetLabelAsSet() {
+		Map<String, String> labels = new HashMap<>();
+		labels.put("k1", "v1,v2,v3");
+
+		PolarisRouterContext routerContext = new PolarisRouterContext();
+		routerContext.putLabels(PolarisRouterContext.ROUTER_LABELS, labels);
+
+		Set<String> resolvedLabels = routerContext.getLabelAsSet("k1");
+
+		Assert.assertEquals(3, resolvedLabels.size());
+		Assert.assertTrue(resolvedLabels.contains("v1"));
+		Assert.assertTrue(resolvedLabels.contains("v2"));
+		Assert.assertTrue(resolvedLabels.contains("v3"));
 	}
 }